Begin translating the image processing instructions from Python to R
Mopy uses its own image processing algorithm to recolour the screenshot. This
translation from Python to R is an attempt to convert that algorithm (for
parity) so that I can render images in the same way that Mopy did.

In the future (and preferrably) I will use functions from magick and imager
instead.